test: split live integration checks by client
